Update
Added a Spawning System
Update Brute Specific Behaviors
Add Sniper Enemy Type
Health System Implemented for Player and Enemies
Add Game Over Functionality

+++ Assets/Scriptable Objects/Shields/ShieldSO.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "ShieldSO", menuName = "Scriptable Objects/ShieldSO")]
 public class ShieldSO : ScriptableObject
 {
-    public int ShieldAmount;
+    public float ShieldAmount;
     public float RechargeRate;
     public GameObject ShieldPrfab;
 }
